python/code_tracing/pyproject.toml
[tool.poetry]
name = "temporalio-samples"
version = "0.1a1"
description = "Temporal.io Python SDK samples"
license = "MIT"
authors = ["Temporal Technologies Inc <sdk@temporal.io>"]
readme = "README.md"
homepage = "https://github.com/temporalio/samples-python"
repository = "https://github.com/temporalio/samples-python"
documentation = "https://docs.temporal.io/docs/python"
packages = [
    { include = "**/*.py", from = "."  }
]

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/temporalio/samples-python/issues"

[tool.poetry.dependencies]
python = "^3.9"
temporalio = {extras = ["opentelemetry"], version = "^1.10.0"}
ddtrace = "^3.2.1"
opentracing = "^2.4.0"
fastapi = {extras = ["standard"], version = "^0.115.11"}
uvicorn = {extras = ["standard"], version = "^0.34.0"}

[tool.poetry.dev-dependencies]
black = "^22.3.0"
isort = "^5.10.1"
mypy = "^1.4.1"
pytest = "^7.1.2"
pytest-asyncio = "^0.18.3"
frozenlist = "^1.4.0"
types-pyyaml = "^6.0.12.20241230"


# All sample-specific dependencies are in optional groups below, named after the
# sample they apply to

[tool.poetry.group.bedrock]
optional = true
[tool.poetry.group.bedrock.dependencies]
boto3 = "^1.34.92"

[tool.poetry.group.dsl]
optional = true
dependencies = { pyyaml = "^6.0.1", types-pyyaml = "^6.0.12", dacite = "^1.8.1" }

[tool.poetry.group.encryption]
optional = true
dependencies = { cryptography = "^38.0.1", aiohttp = "^3.8.1" }

[tool.poetry.group.gevent]
optional = true
dependencies = { gevent = { version = "^23.9.1", python = ">=3.8" } }

[tool.poetry.group.open_telemetry]
optional = true
[tool.poetry.group.open_telemetry.dependencies]
temporalio = { version = "*", extras = ["opentelemetry"] }
opentelemetry-exporter-otlp-proto-grpc = "1.18.0"

[tool.poetry.group.pydantic_converter]
optional = true
dependencies = { pydantic = "^2.10.6" }

[tool.poetry.group.sentry]
optional = true
dependencies = { sentry-sdk = "^1.11.0" }

[tool.poetry.group.trio_async]
optional = true
dependencies = { trio = "^0.28.0", trio-asyncio = "^0.15.0" }

[tool.poe.tasks]
format = [{cmd = "black ."}, {cmd = "isort ."}]
lint = [{cmd = "black --check ."}, {cmd = "isort --check-only ."}, {ref = "lint-types" }]
lint-types = "mypy --check-untyped-defs --namespace-packages ."
test = "pytest"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"

[tool.isort]
profile = "black"
skip_gitignore = true

[tool.mypy]
ignore_missing_imports = true
namespace_packages = true

[[tool.mypy.overrides]]
module = "aiohttp.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "opentelemetry.*"
ignore_errors = true

[tool.poetry.group.cloud_export_to_parquet]
optional = true
[tool.poetry.group.cloud_export_to_parquet.dependencies]
pandas = {version = "^2.2.2",python = ">=3.9,<4.0"}
numpy = {version = "^1.26.0",python = ">=3.9,<3.13"}
pyarrow = "^16.0.0"
boto3 = "^1.34.89"

